Changed indentation and added fallback for bad files
The indentation for the data file format was changed from
three spaces to four spaces, to be more aligned with most development
editors that insert tab mark icons after four spaces.
In case of an error when reading the file (file doesn't exist, etc)
the standard file dialog opens to have the user select another one.