Fixed bug on copy node
Bug when copying a node and then moving the original (not the copy) to a
different parent - the trt data file was not readable any more. This was
fixed. Turned out there were some serious typos in the item's deepcopy
function that caused a reset of data in the original, after the copy
(instead of resetting those data in the copy).