Made selection static during node moves
When a node is moved to a new parent in the current tree, the node
is kept open in the editor after the move. This was different before,
after its move one of the siblings got selected. Keeping the node
open is necessary if you want to move the node down several branches
of a tree.