Fixed problem when trying to create node without selection
Previously, when a tree had no nodes, it was impossible to create any
nodes. New nodes (children, siblings, parents) could only be created
when a node was selected. This was a deadlock after all nodes are
removed.
With this commit, siblings, children or parents can be created when no
selection is present. They will be created as root nodes. With this,
new nodes can be created for empty trees.