Fixed crash on change-parent (again)
When adding a node to a new parent, treetime sometimes crashed.
This was (probably) caused by a problem in the QT Library. Fixed
by explicitly removing and then adding a new parent button (instead
of just updating it). The message
    "QObject::disconnect: Unexpected nullptr parameter"
still appears, but now the software doesn't crash any more. The
explicit removal has (probably) mitigated a race condition (a bit).