Fixed error on node deletion
In the course of debugging I'd put in a dedicated function for
renumbering the root children. This broke deletion of root items
(their siblings didn't get renumbered and the result was a corrupt
data file). Also, the nodes are saved in an ordered way, making
browsing for errors in the data file easier, and speeding up the
load process.