Implemented auto-removal of empty nodes
If the data file contains wrong counting (gaps in the child count),
the item-free nodes used to remain in the tree. They are now removed
and the family is renumbered.
Also, removed the extra widget clearing in treetime.py that had been
put in to prevent a crash. It didn't work anyway.