Fixed sefault crash on parent change
The commit "9c058b38a Speed up of data display"
caused a segfault in the QT library. This crash
is prevented by clearing the whole table area
before re-writing it, and only appears on node
parent moves. Re-introducing this tablewidget.clear()
call fixes it. To mitigate the speed loss, the
function locks the view until the end, so the selection-
changed method is only called once (instead of thrice).