Implemented the first version of time counters
New data field type "timer", with a start and a stop button.
When starting, starts counting the time (in hours, resolution is
four digits after zero - 0.36 seconds). Result is the number of
hours as a floating point number, usable in all tree fields.

Remaining problems: When closing TreeTime while timers are running,
the threads aren't cancelled. This leads to problems after closing.
When loading a file with running timers inside, the auto update of
the tree view is not enabled. Counting works, and when stopping it
shows the correct time, but to see the updates, you must stop and
restart shortly.