Fixed non-saving text when closing or changing trees
When closing TreeTime while an text field with
freshly edited text still has focus, the changes are lost.
The same happened when the tab was changed to another tree.
This is fixed by writing the contents when the focus is lost,
and when the mouse leaves the field. (overriding closeEvent,
destroy() or destroyed() didn't work, for different reasons).