Re-implemented Timer Updates
Since the standard python timers kept crashing when used
in a cascaded way (timer calls timer calls timer), after
1-2 hours in an irregualr way, the timer update function was
re-implemented using one single QTimer that updates all running
timer fields.

Still not done: When closing the software with a running timer,
and then opening it again, the timer does not auto-update any more.

Also still to do: Mark the field values in the GUI of running timers
with an appended "R" or "*" or similar, to indicate they're running.