Implemented export text to clipboard
In order to easily paste tree graphics anywhere, a user may
choose a branch or entire tree, hit the "Text to Clipboard"
button in the Export group, and have unicode graphics (box
drawing elements) in the clipboard. This exports both to the
standard clipboard (ctrl+V), as well as to the selection
(middle most click in Linux).