Made main splitter views adjustable, documentation, minor fix
The main splitter view (data fields on the left, tree on the right) now
has adjustable sizes. You can move the central border with the mouse to
change the width of the data view and the tree view.
Also, fixed the export of the "set" field (was unstyled), also, started
documenting the data format.