Implemented continuous export
User can select an export file (file dialog) and select "continuous
export" before. That file will now be updated once every two seconds,
depending on the selection and on the data. If the continuously exported
file is an html file, there will be a meta tag with a once-a-second
refresh in the file, so a web browser will show the updates immediately.
Using this, you can present/discuss data using treetime for editing and
navigating, and web browsers for viewing.