Continuous export in text field, child layout on export
When inside a text field (editing), and continuos export is
running, the text field contents are saved one per second.
This is in order to not have to leave the field when you want
the html to get updated (i.e. while writing, in "meeting mode").

The saved html child layout was adapted: On limited level export,
nodes that have many children, but these are hidden due to the
level limit setting, will appear next to each other in html/tile
output.