Adapted colours on image export
Adapted the exported colours on the image export. Due to the
white background they must not be too near white, but also
not too vibrant. Equal distances in the colour value spectrum
do not look equal to the eye, and the same goes for the
brightness vs perceived brightness, which is a pain. Looks
okay on my screen now.