Added more user friendly web server message
The message by the web server to the user (now running on...) includes a LAN address
(if available), and a local address, telling them which is which. Also, the name
of the IP isn't saved to the settings, to avoid using it as a file name suggestion
next time something's saved.