Implemented hidden tree fields
If chained calculations are needed but intermediate values not
interesting, intermediate field can be made *hidden* by using
the keyword `hidden` at the end of the field definition.
Hidden fields can be used for calculations in the tree, but
will neither be visible in the GUI, nor in the export.