Fixed shared parameter list problem.
This was a really weird bug. When creating two new tree fields in a row, the parameter
lists were the identical object in the two (changing one changed the other). This was
fixed by not using mutable defaults in the parameter list of the function any more
(changed '[]' to 'False'). Go figure.