First implementation of types
Tree fields and data fields now have types: 'numerical', 'text', 'any'.
Input types are checked, input fields are now only selectable if the
type is correct. Wrong fields are noted with an error message in the
dropdown list (okay, a bit obscure, you actually have to look actively).
Some improvements to follow.
Oh, and the code is not super pretty. It's late and I should actually
be doing other things.